collect: Add /proc/[pid]/sched metrics
Evaluating when a thread is blocked due to block IO requires an
understanding of the time the task spent in the different CPU states,
i.e., sleep (TASK_INTERRUPTIBLE), block (TASK_UNINTERRUPTIBLE) and
on-CPU. Provided the kernel is built with CONFIG_SCHEDSTAT=y, and that
sysctl `kernel.sched_schedstat=1` then the /proc/[pid]/sched file is
written with additional scheduler statistics to be scraped.

This commit adds the code to the metric collector responsible for
scraping the /proc/[pid]/sched file, and persisting the data to a csv
file.

Additionally, the metric collector has been refactored to split the
different application responsibilities between different modules. The
most significant changes are:

* A new `metrics` module is presented. The code to scrape each metric
  subsystem should go within this module. For the moment, there are only
  scheduler metrics.
* A trait is presented to abstract the different metric collectors for
  each subsystem.