collect: Add FIFO process propagation
To collect vfs related metrics, we have to trace the `vfs_read` and
`vfs_write` internal kernel calls.

FIFO files are distinguished from regular files. In the former case, a
thread may hang while reading data from the pipe if there is no data to
be read. For this reason because there is an additional dependency to an
external process that shares the FIFO, the bpftrace program emits an
event indicating there is a new process to be traced.

With regards to regular files, we do not perform process propagation as
simulataneous readers and writers are not considered dependencies from
the filesystem's point-of-view.