collect: Store all bpftrace ipc stat snapshots
Our bpftrace ipc program emits statistics every second. If our program
falls behind, we would like to store all of the snapshots, instead of
calculating an aggregate of those snapshots. This commit adds this new
capability.

Additionally, the program used to compute an aggregate sum of all fifo
statistics. This is no longer performed in the collector in favor of
doing so in the analysis module.