collect: Rename socket file if already exists
To determine the destination file to which data should be written for a
socket, we leverage the global `kfile_socket_map`. However, it is
possible that socket statistics have been persisted before its entry in
the `kfile_socket_map` has been added. This lead to a single socket
producing 2 files with statistics for the same socket: One prior to the
connection entry being added to the global map; another after the
connection entry being added.

In this scenario, when persisting data for the same socket after the
`kfile_socket_map` has its entry, this commit renames the old socket
file, and persists the new data in the same file.