Adapt bpf clone for containers
When a new thread is created in the context of a thread within a
container, the thread identifier returned to the calling parent would be
in the context of the container's namespace. I.e., the returned thread
identifier does not make sense in the context of the host.

Instead of using the value returned from the clone syscall, this commit
changes clone to use bpftrace's builtin `tid`, emitted in the context of
the new thread.